requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via
    #   -r requirements.in
    #   django
build==1.0.3
    # via pip-tools
click==8.1.7
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
defusedxml==0.7.1
    # via
    #   -r requirements.in
    #   odfpy
diff-match-patch==20230430
    # via
    #   -r requirements.in
    #   django-import-export
dj-database-url==2.1.0
    # via -r requirements.in
django==4.2.7
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-import-export
django-import-export==3.3.3
    # via -r requirements.in
et-xmlfile==1.1.0
    # via
    #   -r requirements.in
    #   openpyxl
gunicorn==21.2.0
    # via -r requirements.in
import-export==0.3.1
    # via -r requirements.in
markuppy==1.14
    # via
    #   -r requirements.in
    #   tablib
odfpy==1.4.1
    # via
    #   -r requirements.in
    #   tablib
openpyxl==3.1.2
    # via
    #   -r requirements.in
    #   tablib
packaging==23.2
    # via
    #   build
    #   gunicorn
pillow==10.4.0
    # via -r requirements.in
pip-tools==7.3.0
    # via -r requirements.in
psycopg2==2.9.9
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   tablib
sqlparse==0.4.4
    # via
    #   -r requirements.in
    #   django
tablib[html,ods,xls,xlsx,yaml]==3.5.0
    # via
    #   -r requirements.in
    #   django-import-export
typing-extensions==4.8.0
    # via dj-database-url
tzdata==2023.3
    # via
    #   -r requirements.in
    #   django
wheel==0.41.3
    # via pip-tools
whitenoise==6.6.0
    # via -r requirements.in
xlrd==2.0.1
    # via
    #   -r requirements.in
    #   tablib
xlwt==1.3.0
    # via
    #   -r requirements.in
    #   tablib
python-dotenv
channels==3.0.4
channels_redis
daphne
langchain
mistral
PyPDF2
faiss-cpu
